Minestudio/docs/requirements.txt
numpydoc
linkify-it-py
rich
sphinxext-rediraffe
sphinx-sitemap
sphinx-autoapi>=3.0.0
myst-parser
ablog>=0.11.8
jupyter_sphinx
pandas
plotly
matplotlib
numpy
xarray
sphinx-copybutton
sphinx-design
sphinx-togglebutton
jupyterlite-sphinx
sphinxcontrib-youtube>=1.4.1
sphinx-favicon>=1.0.1
ipykernel
nbsphinx
ipyleaflet
colorama
ipywidgets
graphviz
sphinx_rtd_theme
sphinxcontrib.mermaid
sphinx_copybutton
get_pypi_latest_version
pydata-sphinx-theme
sphinx-multiversion
ray[train]
av==13.1.0
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
setuptools
tqdm
numpy
requests
ipython
typing
gym
gym3
gymnasium
hydra_colorlog
hydra-core>=1.3.2
lmdb
wandb
torch>=2.3.1
lightning
scipy
rich
coloredlogs
daemoniker
lxml
diskcache
albumentations
einops
timm
transformers
Pyro4
xmltodict
pyrender==0.1.25
pyglet
imgui
pyopengl
ray
minecraft_data==3.20.0
absl-py
dm-tree
x_transformers==0.27.1